requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
defusedxml==0.7.1
dj-database-url==3.1.2
Django==6.0.3
django-allauth==65.15.0
django-environ==0.13.0
django-widget-tweaks==1.5.1
gunicorn==25.1.0
idna==3.11
oauthlib==3.3.1
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
